Revert "Xconcat tests; split concatenation tests"
This reverts commit 7c38a58ce94b6e7677e7bbada3ff576474242714.
